refactor: break out override fns from utilities
