+++ src/index.js
@@ -1,7 +1,3 @@
 const customizers = require("./customizers");
 const utilities = require("./utilities");
-
-module.exports = {
-  ...customizers,
-  ...utilities
-};
+const overrides = require("./overrides");
