+++ src/utilities.js
@@ -1,5 +1,3 @@
-const flow = require("lodash.flow");
-
 const getBabelLoader = config => {
   const babelLoaderFilter = rule =>
     rule.loader &&
@@ -23,17 +21,4 @@ const getBabelLoader = config => {
   return babelLoader;
 };
 
-const override = (...plugins) => flow(...plugins.filter(f => f));
-
-// Use this helper to override the webpack dev server settings
-//  it works just like the `override` utility
-const overrideDevServer = (...plugins) => configFunction => (
-  proxy,
-  allowedHost
-) => {
-  const config = configFunction(proxy, allowedHost);
-  const updatedConfig = override(...plugins)(config);
-  return updatedConfig;
-};
-
-module.exports = { getBabelLoader, override, overrideDevServer };
+module.exports = { getBabelLoader };
